requirements.txt
Flask==2.1.2
openai==0.23.0
numpy==1.22.3
python-dotenv==0.20.0
matplotlib==3.5.1